refactor: 자동차 경주 관련 테스트 분리
